Some extra stuff
i's okay if you dont like this change. it's just purely for the aesthetics. basically, I imported the time library and used the time.sleep() function. time.sleep() tells the code to 'sleep' for however long you set it to. like time.sleep(2) -> the code will wait two seconds before displaying the next line of code. it's cool and all, but if you think it's annoying then I'll remove it