terrace update
damn 1055 lines of code. damn.
btw, i feel like we can't have the insanity ending in the actual terrace sequence. it's just because what if the player chooses to jump off the roof in the first run-through? does the game just abruptly end? it's kinda weird when you think about it. if you don't like the stay or leave option, then we can take it out, we'll discuss on zoom later.